move download and unzip to utilities class
